pyproject.toml
[tool.poetry]
name = "war-challenge-computer-vision"
version = "0.1.0"
description = ""
authors = ["Labicom TEAM <pedroserpah@gmail.com>"]
readme = "README.md"
packages = [{include = "war_challenge_computer_vision"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.3"
opencv-python = "^4.8.0.76"
pytesseract = "^0.3.10"
pillow = "^10.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.284"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 115
show-source = true
include = ["**/*.py", "**/*.pyi", "**/pyproject.toml", "*.ipynb"]
select = [
  "E",    # pycodestyle
  "F",    # pyflakes
  "UP",   # pyupgrade
  "N",    # pep8-naming
  "C90",  # mccabe
  "I",    # isort
  "PERF", # perflint
  "NPY",  # numpy
  "PL",   # pylint
  "RUF",  # ruff
]
